Major changes
I've put the header and footer to a seperate folder called static for
easier modification across all pages on the site.
I've started on the login and sign up forms including the connection to
the database. The database connection is currently not stable so that
will have to be fixed.
Added a simple forbidden page on the server that will dedirect all users
who are trying to access places on the webserver they shouldn't be
looking at for example vorur.info/php is forbidden and will redirect
traffic away from the directory and to the forbidden file.